mka: 1.extending Simulation about more accurate granularity (1us) - solved bug 0000022 2.marker modification for TraceLine 3.Rendering3DObject divided into two independent Rendering Pipelines: Render and RenderShadow - solved bug 0000001 4.Updating documentation
git-svn-id: https://localhost:8443/svn/expectance@107 aa805ad0-02f7-422c-84f2-0395ce76bb97

+++ trunk/src/generator/src/framework/core/updatemanager.h
@@ -46,9 +46,10 @@ namespace ft
         
         static UpdateManager* m_instance;
 
-        unsigned int m_lastTick;
+        
+		unsigned long long m_lastTick;
 
-        float m_fpsDuration;
+        double m_fpsDuration;
         int m_fpsFrames;
         int m_fps;
         float m_timeScale;
